+++ patzilla/navigator/services/composite.py
@@ -4,7 +4,7 @@ import logging
 from cornice.service import Service
 from pyramid.httpexceptions import HTTPNotFound, HTTPTemporaryRedirect
 from patzilla.access.drawing import get_drawing_png
-from patzilla.access.generic.pdf import pdf_universal, pdf_universal_multi_zip
+from patzilla.access.generic.pdf import pdf_universal, pdf_ziparchive
 from patzilla.util.date import datetime_iso_filename, now
 
 log = logging.getLogger(__name__)
@@ -47,20 +47,20 @@ def pdf_serve_single(request):
     patent = request.matchdict['patent']
     #parts = request.matchdict['parts']
 
+    # Acquire PDF document.
     data = pdf_universal(patent)
 
-    if 'pdf' in data and data['pdf']:
+    if data.pdf is not None:
         # http://tools.ietf.org/html/rfc6266#section-4.2
         request.response.content_type = 'application/pdf'
         request.response.charset = None
         request.response.headers['Content-Disposition'] = 'inline; filename={0}.pdf'.format(patent)
-        request.response.headers['X-Pdf-Source'] = data['datasource']
-        return data['pdf']
+        request.response.headers['X-Pdf-Source'] = data.datasource
+        return data.pdf
 
-    elif 'meta' in data:
-        meta = data['meta']
-        if 'location' in meta and meta['location']:
-            url = meta['location']
+    elif data.meta:
+        if 'location' in data.meta and data.meta['location']:
+            url = data.meta['location']
             log.info('Redirecting PDF request to "{url}"'.format(url=url))
             raise HTTPTemporaryRedirect(location=url)
 
@@ -72,7 +72,7 @@ def pdf_serve_multi(request):
     patents = patents_raw.split(',')
     patents = [patent.strip() for patent in patents]
 
-    data = pdf_universal_multi_zip(patents)
+    data = pdf_ziparchive(patents)
     zipfilename = 'ip-navigator-collection-pdf_{0}.zip'.format(datetime_iso_filename(now()))
 
     # http://tools.ietf.org/html/rfc6266#section-4.2