+++ patzilla/util/network/browser.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
-# (c) 2017 Andreas Motl, Elmyra UG <andreas.motl@elmyra.de>
+# (c) 2017-2019 Andreas Motl <andreas.motl@ip-tools.org>
 
-regular_user_agent = u'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_7_5) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/49.0.2623.112 Safari/537.36'
+regular_user_agent = u'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_3) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/68.0.3421.0 Safari/537.36'